feat: update ExecutorCoordinator with subscription flow (#1198)
I'm working on adding subscriptions to our federation offering. The flow
of subscriptions is fairly similar to a normal federation execution but
with a few of important differences:

1. There can only be a single subscription field at the root.
2. That subscription field will translate into a subscription call to a
subgraph.
3. We build up a fresh GraphQL response from each response we get back
from that initial subscription.

This PR updates the ExecutorCoordinator to support the subscription flow
as well as the normal flow. I want to return a stream for subscriptions,
ideally without spawning an async task, which needed a lot of
refactoring. Mostly removing references where I could, as they

I needed to do a surprising amount of refactoring to get this working.
Mostly turning references into owned things where it made sense, and
removing the mutable parts of ExecutorCoordinator into local variables
instead.

Most of the changes in this PR are still dead code - my next PR will
hook them up.

Part of GB-5708