feat: expose `execute_stream` from engine-v2 (#1199)
A follow up to #1198 - exposing the changes I made there to consumers of
`engine-v2`. I've not hooked this up in the CLI yet because I want to
get subscriptions working first. This does allow the code to be called
in tests though.

I've added a passing test of queries & mutations via this function just
to confirm.

Part of GB-5708