nix: derivation for the cli (#1212)
This lets anybody with nix installed build the CLI from source with `nix
build github:grafbase/grafbase#cli`, or run it with `nix run
github:grafbase/grafbase#cli`, or build it locally with `nix build
.#cli`, etc. This is also useful documentation on all the inputs
required to build the Grafbase CLI.

This commit also introduces flake-parts to the repo. It is just a
convenient way to organize the derivations that reuses the same module
system any nixos or home-manager user will be familiar with.

One caveat though: the version and URL of the assets.tar.gz file are
hardcoded. Ideally, we would also build these from source, but that can
be done in a separate PR.