fix: concurrent joins in the graphql connector (#1224)
A user on discord is trying to join two GraphQL connectors together. The
type containing the join appears inside a list. When that list has > 1
item, we need to fetch > 1 field from the downstream API. These fetches
get deduplicated into a single query (cool - glad that works) but the
fields in the query don't get aliased correctly.

So you end up with something like:

```
query {
  foo(id: "1") { .. }
  foo(id: "2") { .. }
} 
```

Which is not valid. I fixed this by always aliasing fields in joins
which should solve this problem.

But this uncovered another issue: the graphql connectors deserialisation
wasn't taking aliases into account at all. So even leaving aside joins
any fetches of unnamespaced fields that use aliases would return null.
So I fixed that as well and added a test case.