operation-checks: make sure we handle wrapper types correctly
We were only looking at the outermost wrapper type previously and not
really handling lists, not to speak of nested lists. This commit takes
the wrapper types (! and []) into account on field type changes in
operation checks.

In addition, expose CheckDiagnostic and Severity in public API. It was
an omission.

closes GB-5749