chore: rework federated dev a bit (#1231)
The existing setup in federated dev has the gateway living inside
`Router` with all requests passed in as messages using `RequestSender`.

This works fine at the moment, but I'm currently adding subscriptions
which don't fit into the same request<->response model as well. Each
request requires multiple responses _and_ (at least with websockets)
auth is no necessarily longer tied to a graphql operation - we can auth
connections and every operation after that runs without auth. The
websocket system also needs to observe configuration updates so it can
end subscriptions when they occur.

Updating the router to support all this was getting messy so I've
refactored stuff instead. The `Router` actor has been replaced with
`GatewayNanny` - the only job of which is to combine the config & graph
into a `Gateway` and publish that on a tokio watch channel.

The rest of the system can then use that watch channel to access the
`Gateway` & react to updates.

Part of GB-5715