operation-checks: implement left out argument awareness
In operation checks, we not only care about the arguments queries are
using, but also the arguments they are not using, because they could
become mandatory, either because they go from nullable to nonnull, or
because their default is removed.

This commit makes operation-checks aware of when arguments are left out
so we can detect when a diff on argument default or type could cause
issues. The test cases in the commit should cover all relevant
scenarios.

closes GB-5748