chore: reworked streaming transport code (#1245)
The existing streaming transport code made use of `wait_until` to pump
messages from one place to the other. This was causing me lifetime
problems when integrating into engine-v2, so I've reworked everything to
happen inside the returned `Stream`.

I'm hoping this will work on production as well - we had to be careful
about where the work was done previously so it was definitely inside the
request context. But I think doing the work inside the returned stream
_should_ count as inside the context. I'll test once this is merged
though.