feat: hook subscriptions up to auth (#1252)
Up till now subscriptions weren't authenticated. This implements that,
and adds some tests in the CLI.

Websockets made this a tiny bit awkward - they commonly get
authenticated at the point of connection rather than when an actual
subscription comes in. I've reworked the `Gateway` a bit to handle this
- rather than doing auth inside the `execute` calls, it now exposes an
`authenticate` function that handles auth and returns a `Session` that
exposes the `execute` & `execute_stream` functions.

Also had to fake up an idp server in order to test this in the CLI
tests.

Fixes GB-5801
Fixes GB-5717