chore: make `Gateway::authorize` easier to work with
I added `Gateway::authorize` recently, and made it return an
`engine_v2::Response` on error, as it was the easiest thing to integrate with
all the transports.

However, in the `api` repo the code _really_ wants to take a
`gateway_v2::Response` - which I'd avoided using in `grafbase` because it
doesn't work well with the streaming transports.

This updates the code to just return an `Option`, along with adding some easy
functions for constructing the various error responses.