feat: add websocket urls into the federation config
I'd hard coded this in previous commits so I could do testing without messing
about with configuration - this is a more "proper" implementation.  By default
we use the normal subgraph URL, but users can override with a specific URL to
use for websockets.

I'll do the SDK changes for this in a separate PR.

Part of GB-5711