chore: rework engine-v2 introspection
The primary goal was to remove a bunch of code that is fundamentally useless and broken. I added
a sort of Response builder initially supporting error propagation for
cases when we write data ourselves. But this case actually only happens
for introspection. Everything else receives data from somewhere and
needs to validate it. Furthermore there aren't any possible errors
during introspection, they're either caught during query validation or
when building the schema.

So removing all of that code to build the response manually making it
easier for later changes.

As a bonus it improves introspection speed by ~30%, now it spends most
of its time preparing the operation.