composition: federated graph v2
This PR was prompted by GB-5894, but there are other changes to
FederatedGraph that were discussed with @brabier, so they are bundled
here in a new FederatedGraph version.

Specifically:

- `Value::Float` becomes an f64 to align with the spec.
- Arguments of output fields and fields of input objects are unified
  into InputValueDefinition for code reuse and, more importantly,
  common identifiers for both, so we can treat them uniformly.

We also wanted to represent defaults of arguments but 1. we would need
to check how they are composed, 2. we can add them backwards compatibly
later and this PR is already large enough.

closes GB-5894