composition: federated graph v2 (#1300)
This PR was prompted by GB-5894, but there are other changes to
FederatedGraph that were discussed with @Finistere , so they are bundled
here in a new FederatedGraph version.

Specifically:

- `Value::Float` becomes an f64 to align with the spec.
- Arguments of output fields and fields of input objects are unified
into InputValueDefinition for code reuse and, more importantly, common
identifiers for both, so we can treat them uniformly.
- Directives are treated more uniformly across the schema and stored
separately. That had to happen because we have a new Directive type in
V2 because of the new Value type containing Float(f64). That change
trickled down to all the structs with a Vec<Directive>, and made the
upgrade from V1 to V2 painful. Future changes to directives will be much
smoother with this decoupling.

We also wanted to represent defaults of arguments but 1. we would need
to check how they are composed, 2. we can add them backwards compatibly
later and this PR is already large enough.

closes GB-5894