fix: fields of basic types can have arguments. (#1339)
I added a bit of validation around arguments in #1333 which caused my
tests to fail. Turns out this is because we've been dropping arguments
of the fields basic types rather than putting them into the Registry.

I'm surprised this didn't cause issues elsewhere tbh, expect we're not
being very strict/have other bugs.

Easy to fix though, which this does.