chore: attempt to make gateway_core a bit clearer. (#1348)
I'm trying to implement batching in the various gateway crates, but
finding it very hard to follow some of the code. Mostly because we have
too many things named `Response` that are closely related but not quite
the same (and in places we're working with 2-3 of them in a single line
of code)

This attempts to clean that up a bit:

- Renamed `Executor::Response` to `Executor::StreamingResponse` since
it's actually only used for `execute_stream`.
- Made the error handling for `gateway::Response` a bit clearer - we
were doing a `.into()` that acted like some combination of `.unwrap()`
and `map_err().unwrap_err()`. I found this quite hard to follow
(particularly since go to def doesn't work well for `From` impls) so
switched it about a bit.
- Got rid of `with_additional_headers` from `gateway_core::Response` so
that it was entirely functions for creating a response.
- Renamed `gateway_core::Response` to `ConstructableResponse` since it
is now entirely concerned with the construction of responses.
- Got rid of some `Deref`/`DerefMut` impls that barely served any
purpose.

I think this makes things a little easier to follow - though open to
discussion if anyone disagrees (or has alternative suggestions).