feat: graphql batching in cli standalone graphs (#1354)
Implements [this RFC][1] for standalone graphs in the CLI.

The RFC is pretty silent on whether requests should be run in parallel
or in series. Even the implementations linked to from the RFC can't
agree: half of them do things in parallel, half in series. I've settled
on doing things serially for now as it's easy and definitely won't break
anything. Can revisit later if we need to.

Part of GB-5931

[1]: https://github.com/graphql/graphql-over-http/blob/main/rfcs/Batching.md