engine-v2-schema: remove inaccessible items
Since they are not part of the API of the engine, they should be excluded
as early as possible. They can fall through the cracks if they are
included in the schema, so we remove them at conversion time.

The approach taken is to remove inaccessible output fields, input value
definitions, union members and enum values. Since inaccessible types
cause the fields returning them to also be inaccessible (composition
rule), that automatically deals with inaccessible types.

closes GB-5949