feat: Validate all inputs part 1
This PR only includes the engine-v2-schema changes which were
necessary to re-use `RawInputValues` for operation input values and
make them pleasant to work with.

With the next PR it allows us to fully isolate:
- all the GraphQL-specific input processing: default values, undefined variables, coercion
- schema names from connector/data sources names. Virtually we can
  trivially rename anything now as we keep track of input value
  definition ids and enum value ids. And executors have everything
  automatically renamed by the provided SchemaWalker.

Input values, which will be accessed through a `PlanInputValue`, can be
used through:
- Serialize to serialize the data immediately
- Deserializer to map arguments to a custom type
- Display to format the value with GraphQL syntax
- InputValue<'_> which provides a friendlier format than RawInputValue
  which also abstracts away the differences between schema default
  values and operation input values.

In all cases, the values have always already been coerced and validated.