cli: new command for trusted document submission
The command is called `trust` and lives at the top level, but we can
iterate if that turns out not to be self-explanatory enough.

The command is hidden for now, since the feature isn't complete in the
engine and infrastructure yet, and not documented.

It takes a branch reference, a client name and a manifest file that can
be either in Relay or Apollo Client format. We can easily add more
manifest formats as the appetite arises. We can even imagine using the
Grafbase CLI to generate manifests directly.

closes GB-6015