fix: respect `.gitignore` in `grafbase deploy` (#1394)
A user was running into issues with `grafbase deploy` yesterday, where
deployments failed with a bizzare error very similar to [this one][1].

We eventually tracked this down to files in the users `.direnv` or
`.devenv` folders - based on the tar-rs issue _probably_ hard links of
some description. While it would be good to not choke on hardlinks, I'd
argue we shouldn't have been archiving these files in the first place:
they're clearly meant to stay local to the users machine.

So this PR switches `walkdir` for `ignore`, which will respect
`.gitignore` files etc, providing users with a sensible (and in most
cases zero effort) way to exclude files from the archive we build on
`grafbase deploy`. I think this is a sensible default.

This _could_ have ramifications for anyone doing a `grafbase deploy`
after some kind of local build step, though I don't know how likely that
is really. I suggest we try it and see - if anyone complains we can
revisit, maybe adding a CLI flag to change the behaviour.

Fixes GB-6107

[1]: https://github.com/alexcrichton/tar-rs/issues/313