fix: issues with overlapping fragments in engine-v1 (#1405)
`engine-v1` had an issue when a query came in with overlapping
fragments, e.g.:

```graphql
query {
  foo {
     ... {
        bar {
           baz
        }
     }
     ... {
        bar {
           blorp
        }
     }
  }
}
```

The `foo` object has two different instances of its field `bar` and when
inserting these into the response we were overwriting one with the
other. This meant that the response would only have `blorp` in it and
not `baz`. This isn't likely to be an issue for handcrafted queries, but
is very likely for anyone using Relay - as it very much encourages you
to have overlapping fragments.

The fix I've attempted is just to do a nested merge on duplicate field
names when populating the response. This isn't quite what the spec says
you should do: I think there's meant to be a preprocessing step where
you collect all the fields from each nested fragment into a set _before_
you resolve them. But that'd be a much more drastic change I think so
this'll hopefully do for now.

Fixes GB-6106.