fix: coerce enums when handling join variables (for real this time) (#1451)
This is a fixed version of #1488 - I hadn't noticed a test failure and
somehow it got merged anyway. You can skip over the first commit if you
want, it's only the second & third that are new. Though it's not a big
PR in the first place.

The issue with #1488 was that I was only processing joined variables on
the current field before resolving, ignoring any sub-fields. This
doesn't work with any resolver that looks at the whole sub-query (e.g.
the GraphQL connector). I've updated it to do replacement across the
whole query prior to running any resolvers.

Also added a more thorough test of this issue, since the one I'd written
wasn't great.