feat(tracing): Add query & operation type for subgraph requests
In order to log the query we're sending to the subgraph I'm replacing
all input values in the query string with variables. I also realized how
much allocations I was doing in the query building, so improved that.

I've also simplified the query generation for named fragments, instead
of trying to keep them I just transform them into inline ones. And I'm
only keeping type conditions which are useful. There is little downside
except probably a bit bigger query strings.

The operation type is also now better handled. the `PlanWalker` doesn't
expose `operation` anymore. It was the root operation and was easily
confused with the subgraph graphql operation. As subscription executors
also relied on that, I've also changed the coordinator/subscription
executor interface.