fix: a non-deterministic test
I'd fooled myself into thinking the order that joins were executed was
determinstic - it is not.  This meant that the query snapshot could appear in
one of two orders, making the test flaky.

Switched it so there's only a single fetch needed for the join, which solves
the problem for now.