fix: error handling in graphql connector (#1465)
The GraphQL connector uses data loaders to group concurrent queries into
a single query that can be sent to the downstream server. The handling
of errors for this case was extremely naive: all the errors from the
response were just cloned and sent back to the individual ungrouped
resolvers, each of which would then record all the errors in the
context.

This is particularly problematic when combined with joins on list
fields: the GraphQL connector will join every subquery on the list
elements into a single query. If that response contains errors on all
its field, then you end up with N^2 errors in the GraphQL response. In
one case this seems to have led to ~70MB responses being sent back to
Pathfinder.

I've updated the code to be smarter about errors:

1. Deduplicating errors before sending back to resolvers.
2. Rewriting the path inside errors so its correct in the context of our
API.
3. More thorough tests of error handling