feat: add argument support to the SDK (#1468)
I'm adding argument support to joins at the moment, and while planning
how to do this I noticed the SDK has no way to add arguments to
arbitrary fields. It lets you add arguments to fields of the root query
type but not anywhere else.

This PR fixes that by adding `arguments` functions to the
`JoinDefinition` & `ResolverDefintion` types. These are I _think_ the
only places that arguments make sense at the moment, though I'm open to
discussing if anyone thinks a different approach would be better.

I've also added an `allArguments` property to most of the types involved
in specifying field types, even those that might not have their own
arguments. I needed to do this because e.g. `JoinDefinition` can be
wrapped by `AuthDefinition` - so we need to make sure we ask every
object in the fields chain of wrappers to make sure we don't miss any
arguments.