composition: new federation auth directives, FederatedGraphV3 (#1463)
The original purpose of this PR is to support the authentication
directives introduced in 2.5 and 2.6: `@policy`, `@requiresScopes` and
`@authenticated`. That cascaded into a larger pull request because
adding new directives to the `Directives` enum means a new
`FederatedGraph` version, and we had a few changes pending for the next
version.

The largest amount of changes come from

- the move from FieldType as interned value to an inline, small value.
This is a simplification in many places, but it required moving the
wrapping type representation of engine-v2 to its own crate. I am open to
alternative solutions.
- the move from separate `object_fields` and `interface_fields` vecs to
inline IdRanges for interface and object fields. That introduces one big
constraint: the fields for a given object/interface have to be
contiguous in the `fields` vec. That new constraint demanded changes
from composition all the way to introspection in engine v2.

Composition also now sorts fields by name.

closes GB-6125
closes GB-6124
closes GB-6133