feat: arguments on join fields (#1471)
A customer wants to forward an argument on a join field into the query
that we use to perform the join. This adds support for that, using the
same variable syntax we already use.

Any arguments on the join field are automatically available under
`$argumentName`. Currently arguments will always shadow fields of the
parent object. So if there's a `foo` argument then `$foo` will _always_
refer to that argument, and any field on the parent type named `foo`
will be inaccessible. The alternative to this would have been to use
some special syntax to differentiate fields & arguments, but that seemed
like a lot more work, and I'm not sure how likely it is for users to
need that functionality. We can always revisit later.

I'm not doing much validation on these arguments at parsing time
currently. I'll do that in a follow up PR, just wanted to unblock the
customer as quick as possible.

Fixes GB-6304