chore(engine-v2): Split input values into different structs
The current input value merged various things into a single struct with
generics. That was a mistake. So I'm splitting them into three:

- DefaultInputValue: provided by Schema, the current InputValue stuff.
  Will remove unused code later.
- VariableInputValue: InputValue coming from a variable (so always a ConstValue).
- QueryInputValue: InputValue coming from the query which can reference
  a variable.

default values for variables are transformed into QueryInputValue and
not VariableInputValue because the distinction also comes from the fact
than one is coming from the `engine_parser` and other from `Request`.
Currently both use `ConstValue`, but that's something I want to
eventually change. The `engine::Variables` isn't great.

The serde/Into<InputValue> stuff is mostly copy-pasted.

From now on the `OperationWalker` always includes the variables.
Initially I wanted to create the `OperationPlan` before taking any
variables into account to make it entirely cacheable. But it's just a
lot of work to do this correctly with `@include/@skip`. So later I'll
just cache the `OperationPlan` if it didn't rely on variables.

This split will also allow me to simplify the input values in the schema
as I need to add some logic on those to for `@requires` with arguments.