graphql-federated-graph: fix missing-Query case and upgrade path in FederatedGraphV3 (#1478)
This PR should fix the end-to-end test failures due to FederatedGraphV3.
It fixes two problems that lead to federated graphs without `__type` and
`__schema` fields on the root query type, leading to an assertion
failure in introspection.

1. The actual cause for the e2e test failures:
`FederatedGraph::from_sdl()`. I assumed that we only use this with
schemas produced by composition and then `FederatedGraph::render_sdl()`,
but some test cases have an empty root query type. That made the
provision to add `__type` and `__schema` in `from_sdl()` not trigger,
and the fields were missing as a result. Thanks @obmarg for figuring
this out.

To make sure that composition never produces an empty query type, the PR
also contains a new composition test. We were only testing that a
missing query type is rejected, we now test the empty query type case.

3. Another potential source of failures in the FederatedGraph version
update logic: we weren't adding the fields when upgrading from v2 to v3.
This PR contains a fix and test.

---------

Co-authored-by: Yoav Lavi <yoav@yoavlavi.com>