federated-graph: new export to generate final API schema from FederatedGraph (#1522)
The purpose is for us to stop relying on introspection to generate these
schemas. The implementation reuses bits of the existing `render_sdl()`
(which is renamed to `render_federated_sdl()` for clarity), but mostly,
it is new. In this commit, we modify the composition test setup to reuse
the test schemas we already have there.

closes GB-5666