chore: only use `codegen-units=1` for lambda releases (#1561)
We recently added `codegen-units=1` to the release build config. This
was aimed at reducing the size of the binaries that we ship for lambdas,
and it was successful in that regard. It slows release builds down quite
significantly though, which has had an impact on CI times, as you can
see on this graph.

![CleanShot 2024-04-05 at 12 01
13@2x](https://github.com/grafbase/grafbase/assets/556490/c142187c-b067-4ed9-a1a0-0066ce27acb4)

That graph shows a day or two between the merge & a noticeable increase
in CI times, but the merge was on friday and I assume it wasn't until
the monday that every other piece of work had rebased onto `main` and
pulled in the change.

This PR makes two changes:

- Creates a new cargo profile `lambda` that with `codegen-units=1` and
updates CI to use this profile for building the lambda. This shaved 5-6
minutes off.
- I then updated the lambda build to only use this profile when doing a
tagged release build, which shaved another ~2:30 off. Although I'm open
to discussions if we'd rather not do this bit.