chore: report test results to datadog (#1564)
We upload junit files to datadog in the API repo. I fixed some of this
in https://github.com/grafbase/api/pull/3459 but noticed that we weren't
doing similar in this repo. It can be useful to track failures, slow
tests, flaky tests etc.

So, this PR adds the upload code to this repo as well.