feat: allow tests to define interfaces (#1579)
Back in the days of `@model`, grafbase didn't support interfaces. On
adding the graphql connector, grafbase started having to support
interfaces.

Hugo is currently trying to test some stuff around caching & interfaces,
but he ran into difficulties testing, because `parser-sdl` still just
ignores interfaces. It's quite easy to fix this though, so that's what
I've done.

At some point we'll probably need to do the same for union types as
well, but that's for the future. Also this doesn't techncially let users
do interfaces because I've not updated the SDK, but we can do that if we
want.