fix: consider nested variables in join string
I'd obviously forgotten to update the bit of code that finds variables in the
join stirng to handle variables on nested selections.  This meant that the
validation code for arguments considered any arguments on nested fields as
"unused" even though they clearly were.  Would have affected the implict
requires we put on these fields as well, which I imagine nobody would notice,
because that barely works with half of our connectors anyway.