fix: introspect when string literal ends in a double quote (#1601)
An API we're working with contains the following directive:

```
@deprecated(reason: "Use \"urlSite\"")
```

When we run `grafbase introspect` we take whatever output we have and
run it through `apollo_encoder` to prettify it. Unfortunately
`apollo_encoder` naively makes everything a block string, leading to
this output:

```
@deprecated(reason: """Use "urlSite"""")
```

Note the 4 quotes on the end there. I don't think this is valid GraphQL,
I can't see any way this should parse based on the spec.

I tried to fix this by upgrading `apollo_encoder` but it turns out it's
not even a thing anymore - the functionality has been rolled up into
`apollo_compiler`. I tried using this and it does fix the issue, but the
output from `apollo_compiler` isn't particularly prettified:

```
newsletterSubscribe(email: String!, referer: String!, type: String!, siteOrigin: String!, sourcePage: String!, templateName: String!, ipCountry: String!, tpUuid: String!, ip: String): Boolean!
```

So, I've switched it out for `cynic-parser` which contains an actual
pretty printer. My aim with that pretty printing is to replicate the
output from prettier. It's not _quite_ there yet, but it is pretty close
and will get there sooner or later. For example the above field is
formatted correctly like this:

```
newsletterSubscribe(
  email: String!
  referer: String!
  type: String!
  siteOrigin: String!
  sourcePage: String!
  templateName: String!
  ipCountry: String!
  tpUuid: String!
  ip: String
): Boolean!
```