cli: return early with an error on missing schema in grafbase check
If you run grafbase check without the --schema argument, the CLI will
wait for a schema from stdin, giving the impression it is hanging. Similar
to what we already did for `grafbase publish`, this commit makes the
CLI exit early in this situation by detecting if stdout is a tty.

closes GH 6494