engine-v1: introspect data sources concurrently (#1609)
Introspection of data sources gets noticeably slow once you use enough
connectors that require introspection when starting `grafbase dev`
(openapi, postgres, graphql). We explicitly did not introspect
concurrently until now because of memory usage concerns. So the solution
here is a compromise:

- No multithreading. The CPU/memory bound work will all happen in the
same thread, and it will run from start to end for each connector
sequentially because that work does not yield. Only the IO is somewhat
more concurrent.
- `FuturesOrdered` because although the order shouldn't matter, we
should keep the same deterministic behaviour we had if we can.

If we wanted to parallelize more aggressively, potential candidates
would be a thread per datasource or using tokio's spawn_blocking() for
CPU work, but we would then expect memory usage to go up further, and it
is more intrusive modification in parser-sdl. The solution proposed in
this PR has the advantage of touching a single function and using
dependencies we already have.

The benchmark included here, on main:

```
Benchmarking parse_sdl() with IO/parse_sdl: Collecting 15 samples in estimated 31.088 s (15 iterations
parse_sdl() with IO/parse_sdl                                                                         
                        time:   [2.0002 s 2.1014 s 2.2093 s]                         
```

On this branch:

```
Benchmarking parse_sdl() with IO/parse_sdl: Collecting 15 samples in estimated 14.234 s (15 iterations
parse_sdl() with IO/parse_sdl
                        time:   [962.22 ms 1.0299 s 1.0966 s]
                        change: [-55.071% -50.990% -46.499%] (p = 0.00 < 0.05)
                        Performance has improved.
```

The high water mark in memory usage is pretty consistently around 99.7MB
on main, and 100.7MB on this branch, so the difference doesn't appear to
be significant.

As another data point, in a customer setup I reproduced locally,
`parse_sdl()` went from 8.5 to 2.5 seconds.

closes GB-6504